requirements.txt
arxiv==2.4.1
python-dotenv==1.2.1
anthropic==0.86.0
